analytics_service/requirements.txt
redis==5.2.1
psycopg2-binary==2.9.10
numpy==2.2.1
SQLAlchemy==2.0.36
Flask==3.1.0
Flask-Cors==3.0.10
prometheus-client==0.21.1
pymongo==4.11.1
scikit-learn==1.5.2
waitress==3.0.2